sparse matrix BB is not assembled with Hermetian transpose of BB_U
The assembled sparse matrix (BB) is not created with the Hermitian transpose of BB_U. This leads to significant performance differences for higher orders compared to the Matlab implementation. The used method .transpose() just changes matrix dimensions, but does not change the sign of the matrix's imaginary part.  The operation analogous to Matlab's Hermetian transpose operation BB_U' is BB_U.getH().